LocalAI-Photos/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
pillow==11.0.0
pyyaml==6.0.2
watchdog==4.0.0
faiss-cpu==1.7.4
sentence-transformers==3.2.1
open-clip-torch==2.24.0
torch==2.2.2
torchvision==0.17.2
deepface==0.0.93
easyocr==1.7.2
opencv-python-headless==4.10.0.84
ultralytics==8.0.196
hdbscan==0.8.38
numpy==1.26.4
python-multipart==0.0.9
